Add tabs component
Update Projects page
