+++ packages/client-core/src/admin2/components/project/index.tsx
@@ -23,21 +23,31 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { useHookstate } from '@etherealengine/hyperflux'
+import Tabs from '@etherealengine/ui/src/primitives/tailwind/Tabs'
 import React from 'react'
-import ProjectTab from './ProjectTab'
-import BuildStatusTab from './build-status/BuildStatusTab'
+import { useTranslation } from 'react-i18next'
+import ProjectTable from './ProjectTable'
+import ProjectTopMenu from './ProjectTopMenu'
+import BuildStatusTable from './build-status/BuildStatusTable'
 
 export default function AdminProject() {
-  const selectedTab = useHookstate<'projects' | 'buildStatus'>('projects')
+  const { t } = useTranslation()
 
   return (
-    <>
-      {selectedTab.value === 'projects' ? (
-        <ProjectTab selectedTab={selectedTab} />
-      ) : (
-        <BuildStatusTab selectedTab={selectedTab} />
-      )}
-    </>
+    <Tabs
+      tabsData={[
+        {
+          title: t('admin:components.project.project'),
+          tabLabel: t('admin:components.common.all'),
+          rightComponent: <ProjectTopMenu />,
+          bottomComponent: <ProjectTable />
+        },
+        {
+          title: t('admin:components.buildStatus.buildStatus'),
+          tabLabel: t('admin:components.project.buildStatus'),
+          bottomComponent: <BuildStatusTable />
+        }
+      ]}
+    />
   )
 }
